move tests back to where it belongs
